+++ packages/actionsflow-core/src/index.ts
@@ -16,6 +16,7 @@ export {
   getExpressionResult,
   getStringFunctionResult,
   getThirdPartyTrigger,
+  getGlobalThirdPartyTrigger,
   getLocalTrigger,
   isPromise,
   getRawTriggers,