+++ packages/actionsflow/types/index.d.ts
@@ -3,3 +3,5 @@ declare module "string.prototype.matchall";
 declare module "@actionsflow/trigger-typeform";
 declare module "@actionsflow/trigger-google_form";
 declare module "@actionsflow/trigger-trello";
+declare module "exec-sh";
+declare module "cron";